left click show app on macos
